Added vins to impacted address and input check for req

+++ app/controllers/blacklist.ts
@@ -19,42 +19,58 @@ module.exports = {
         });
     },
     addBlacklist : function(req: any, res: any) {
-        try {
-            bitcoin.address.toOutputScript(req.body.address, network);
-        } catch (e) {
-            const errorType = "Invalid address";
+        if (req.body && req.body.address) {
+            try {
+                bitcoin.address.toOutputScript(req.body.address, network);
+            } catch (e) {
+                const errorType = "Invalid address";
+                const error = new Error(errorType);
+                error.status = 400;
+                res.send(error);
+                return;
+            }
+            const blacklistEntry = new Blacklist(req.body);
+            blacklistEntry.save((err: any) => {
+                if (err) {
+                    res.send(err);
+                } else {
+                    res.send(blacklistEntry);
+                }
+            });
+        } else {
+            const errorType = "No address was included in the request"
             const error = new Error(errorType);
             error.status = 400;
             res.send(error);
             return;
         }
-        const blacklistEntry = new Blacklist(req.body);
-        blacklistEntry.save((err: any) => {
-            if (err) {
-                res.send(err);
-            } else {
-                res.send(blacklistEntry);
-            }
-        });
     },
 
     deleteBlacklist : function(req: any, res: any) {
-        try {
-            bitcoin.address.toOutputScript(req.body.address, network);
-        } catch (e) {
-            const errorType = "Invalid address";
+        if (req.body && req.body.address) {
+            try {
+                bitcoin.address.toOutputScript(req.body.address, network);
+            } catch (e) {
+                const errorType = "Invalid address";
+                const error = new Error(errorType);
+                error.status = 400;
+                res.send(error);
+                return;
+            }
+            const blacklistEntry = Blacklist.deleteOne(req.body, (err: any) => {
+                if (err) {
+                    res.send(err);
+                } else {
+                    res.send("Blacklist entry deleted from database");
+                }
+            });
+        } else {
+            const errorType = "No address was included in the request"
             const error = new Error(errorType);
             error.status = 400;
             res.send(error);
             return;
         }
-        const blacklistEntry = Blacklist.deleteOne(req.body, (err: any) => {
-            if (err) {
-                res.send(err);
-            } else {
-                res.send("Blacklist entry deleted from database");
-            }
-        });
     }
 }
